requirements.txt
fastapi
uvicorn
pydantic
openai
pypdf
python-docx
pytesseract
Pillow
pandas
pdf2image
PyGithub
pytest
httpx
